comentarios corrigidos e implemetacoes finalizadas

+++ Calculadora/Operacoes.cs
@@ -3,10 +3,10 @@ namespace Calculadora
 {
     public class Operacoes
     {
-        //Alterado int para long, erro na operação 2147483647 + 2 necessita de tipo long
-        public long valorA {get;set;}
+        //Alterado int para decimal, erro na operação 2147483647 + 2 é causado pelo tipo int. Tipo decimal é mais coerente para a calculadora, correspondendo também ao tipo já definido ao resultado.
+        public decimal valorA {get;set;}
         public char operador { get; set; }
-        public long valorB { get; set; }
+        public decimal valorB { get; set; }
         public decimal resultado { get; set; }
 
     }
